Switch to criterion for benchmarks
They don't require nightly to run and use statistics for more accurate
benchmarks.
