+++ .github/workflows/ci.yml
@@ -33,7 +33,7 @@ jobs:
     strategy:
       matrix:
         rust:
-          - 1.41.1 # MSRV (minimum supported Rust version)
+          - 1.46.0 # MSRV (minimum supported Rust version)
           - stable
           - beta
     steps:
