+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html),
 with the exception that 0.x versions can break between minor versions.
 
+## Unreleased
+### Changed
+- Bump MSRV (minimum supported Rust version) to 1.46
+
 ## [0.7.0] - 2021-05-18
 ### Changed
 - URLs that have a quote character like `'` or `"` before them will stop
